Updated swagger spec and addressed review comments
